# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "Description of Task Manager API",
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Вход",
                "operationId": "ce5ef90f2ad7d4ab966397fd8ddd231d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Выход",
                "operationId": "11d76349d4369c54169628716b62f9a9",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "operationId": "0e2da547a5af550417e8666500022fbf",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/sanctum/csrf-cookie": {
            "get": {
                "tags": [
                    "csrf"
                ],
                "summary": "Метод для записи csrf куки в браузер",
                "operationId": "a2162b0e12e57cbec91259cc3d187526",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Показ информации о текущем авторизованном пользователе",
                "operationId": "ede45b19680d6fb1319013a741a05734",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/ProfileResource"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Показ задач авторизованного пользователя",
                "operationId": "d65f8818ec696dad558b47939034f1c5",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "title",
                            "nullable": true
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "low,high",
                            "nullable": true
                        }
                    },
                    {
                        "name": "deadline_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-01-01 21:00:00",
                            "nullable": true
                        }
                    },
                    {
                        "name": "deadline_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-01-01 21:00:00",
                            "nullable": true
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "deadline",
                                "priority"
                            ],
                            "example": "title",
                            "nullable": true
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc",
                            "nullable": true
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "example": false,
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task": {
            "post": {
                "tags": [
                    "profile"
                ],
                "summary": "Создание задачи для авторизованного пользователя",
                "operationId": "fc51dc8c4d8726cf188cd0a95bb8c992",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Созданная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/{id}": {
            "put": {
                "tags": [
                    "profile"
                ],
                "summary": "Редактирование задачи авторизованного пользователя",
                "operationId": "02617bdd4f940435e1a575f1aeb43bb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Созданная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/{id}/toggle": {
            "patch": {
                "tags": [
                    "profile"
                ],
                "summary": "Редактирование задачи авторизованного пользователя",
                "operationId": "abc557d6ba991b0912677249b274a25b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PriorityTaskEnum": {
                "description": "low: Минимальный, medium: Средний, high: Наивысший",
                "type": "string",
                "enum": [
                    "low",
                    "medium",
                    "high"
                ],
                "x-enum-varnames": [
                    "LOW",
                    "MEDIUM",
                    "HIGH"
                ]
            },
            "LoginRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test@test.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "RegistrationRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test@test.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "TaskCreateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01 21:00:00"
                    },
                    "priority": {
                        "$ref": "#/components/schemas/PriorityTaskEnum"
                    }
                },
                "type": "object"
            },
            "TaskFilterRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "title",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "example": "low,high",
                        "nullable": true
                    },
                    "deadline_from": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01 21:00:00",
                        "nullable": true
                    },
                    "deadline_to": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01 21:00:00",
                        "nullable": true
                    },
                    "sort_by": {
                        "type": "string",
                        "enum": [
                            "title",
                            "deadline"
                        ],
                        "example": "title",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "example": "asc",
                        "nullable": true
                    },
                    "active_only": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProfileResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test@test.com"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01 21:00:00"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "priority": {
                        "type": "string",
                        "example": "low"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "auth",
            "description": "auth"
        },
        {
            "name": "csrf",
            "description": "csrf"
        },
        {
            "name": "profile",
            "description": "profile"
        }
    ]
}